move map2alm, alm2map to new file
